Credorax: Enable selecting a processor
Adds gateway specific fields to the Credorax gateway to enable the user
to select a different processor, including adding a unit test and a
(failing) remote test.

Gateway specific fields added:
 - R1: processor name
 - R2: processor merchant ID

CE-35

Unit:
23 tests, 106 assertions, 0 failures, 0 errors, 0 pendings, 0 omissions,
0 notifications
100% passed

Remote:
26 tests, 37 assertions, 21 failures, 0 errors, 0 pendings, 0 omissions,
0 notifications
19.2308% passed

The Credorax gateway integration was recently updated, and now many of
the remote tests that were previously passing are failing. It is
currently unknown why these tests are all failing. We have reached
out to Credorax to inquire about this new behavior.